requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
django~=1.8.0
django-extensions==1.5.7
djangorestframework==3.2.3
dj-database-url==0.4.1
google-api-python-client==1.5.2
gunicorn==19.3.0
python-dotenv==0.1.3
pyyaml==3.11
hashids==1.1.0
psycopg2==2.6.1
two1==3.8.0
